% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyappDeployment.R
\docType{class}
\name{ShinyappDeployment-class}
\alias{ShinyappDeployment}
\alias{ShinyappDeployment-class}
\title{Extention of keboola.r.docker.application 
Used to provide an interface between KBC and shinyapps.io}
\description{
Extention of keboola.r.docker.application 
Used to provide an interface between KBC and shinyapps.io
}
\section{Fields}{

\describe{
\item{\code{action}}{A string from the configuration that tells the app what to do. 
The only possible action right now is list.  terminate and deploy are run asynchronously via the default run action.}

\item{\code{params}}{A list of parameters found in the configuration}
}}
\section{Methods}{

\describe{
\item{\code{run()}}{Main application function, this should be overriden in child class.
\subsection{Return Value}{NULL}}
}}

